3차 리팩토링
