+++ Data/GAME_VIEW/OBJECT/view.py
