+++ config.ini
@@ -1,0 +1,11 @@
+[system]
+is_new = False
+color_weakness_mode = FALSE
+screen_width = 1280
+screen_height = 720
+uno = pygame.K_u
+left_move = pygame.K_RIGHT
+right_move = pygame.K_LEFT
+select = pygame.K_RETURN
+draw = pygame.K_z
+
